requirements.txt
# Core dependencies
streamlit>=1.28.0
plotly>=5.17.0
pandas>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0
gql>=3.4.0

# Testing dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.1

# Optional dependencies
streamlit-aggrid
openpyxl
matplotlib
Pillow
